fix: stockCode UUID->String, add: 주식 예측 시 theme card 발급

+++ src/main/java/com/example/legendfive/overall/dto/ThemeDto.java
@@ -20,7 +20,7 @@ public class ThemeDto {
     public static class ThemeCardListResponseDto{
         private String themeName;
         private LocalDate createdAt;
-        private Long themeCount;
+        private int themeCount;
         private String userNickname;
     }
 }
